Separate usecase package
