+++ domain/usecase/auth/error.go
