+++ domain/usecase/auth/service.go
@@ -2,8 +2,8 @@ package auth
 
 import (
 	"github.com/gghcode/apas-todo-apiserver/config"
-	"github.com/gghcode/apas-todo-apiserver/domain/security"
-	"github.com/gghcode/apas-todo-apiserver/domain/user"
+	"github.com/gghcode/apas-todo-apiserver/domain/usecase/security"
+	"github.com/gghcode/apas-todo-apiserver/domain/usecase/user"
 )
 
 type (
