+++ infrastructure/jwt/token_handlers.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 
 	"github.com/gghcode/apas-todo-apiserver/config"
-	"github.com/gghcode/apas-todo-apiserver/domain/auth"
+	"github.com/gghcode/apas-todo-apiserver/domain/usecase/auth"
 	"github.com/gghcode/apas-todo-apiserver/web/middleware"
 
 	"github.com/dgrijalva/jwt-go"
