+++ infrastructure/repository/user_repository.go
@@ -4,7 +4,7 @@ import (
 	"time"
 
 	"github.com/gghcode/apas-todo-apiserver/db"
-	"github.com/gghcode/apas-todo-apiserver/domain/user"
+	"github.com/gghcode/apas-todo-apiserver/domain/usecase/user"
 	"github.com/gghcode/apas-todo-apiserver/infrastructure/model"
 	"github.com/jinzhu/gorm"
 	pg "github.com/lib/pq"
