+++ web/api/todo/validators.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"errors"
 
-	"github.com/gghcode/apas-todo-apiserver/domain/todo"
+	"github.com/gghcode/apas-todo-apiserver/domain/usecase/todo"
 	"github.com/gghcode/apas-todo-apiserver/web/api"
 	"github.com/gghcode/apas-todo-apiserver/web/middleware"
 	"github.com/gin-gonic/gin"
