+++ wire_gen.go
@@ -8,10 +8,10 @@ package main
 import (
 	"github.com/gghcode/apas-todo-apiserver/config"
 	"github.com/gghcode/apas-todo-apiserver/db"
-	"github.com/gghcode/apas-todo-apiserver/domain/app"
-	"github.com/gghcode/apas-todo-apiserver/domain/auth"
-	"github.com/gghcode/apas-todo-apiserver/domain/todo"
-	"github.com/gghcode/apas-todo-apiserver/domain/user"
+	"github.com/gghcode/apas-todo-apiserver/domain/usecase/app"
+	"github.com/gghcode/apas-todo-apiserver/domain/usecase/auth"
+	"github.com/gghcode/apas-todo-apiserver/domain/usecase/todo"
+	"github.com/gghcode/apas-todo-apiserver/domain/usecase/user"
 	"github.com/gghcode/apas-todo-apiserver/infrastructure/file"
 	"github.com/gghcode/apas-todo-apiserver/infrastructure/jwt"
 	"github.com/gghcode/apas-todo-apiserver/infrastructure/repository"
